Show ongoing visits in client
move models common between client and server into util.models
and add pydantic as a client dependency (server already has it as
transitive)
