+++ requirements_dev.txt
@@ -3,6 +3,7 @@ importlib_resources==5.6.0; python_version<'3.9'
 ispyb==6.10.0
 jinja2==3.1.1
 procrunner==2.3.3
+pydantic==1.9.0
 pytest-cov==3.0.0
 pytest==7.1.1
 rich==12.0.1
