+++ setup.cfg
@@ -19,6 +19,7 @@ project_urls =
 [options]
 include_package_data = True
 install_requires =
+    pydantic
     requests
     rich
 packages = find:
